Bugfix in VehicleManager/NavigationManager.

+++ SmartRaceCar/Modules/racecarbackend/src/main/java/be/uantwerpen/fti/ds/sc/racecarbackend/MapManager.java
@@ -216,10 +216,10 @@ public class MapManager implements MQTTListener
 					break;
 				case "gangV":
 					this.log.info("Loading wayPoints for " + this.currentMap);
-					this.wayPoints.put((long) 46, new WayPoint(46, (float) -6.1, (float) -28.78, (float) 0.73, (float) 0.69));
-					this.wayPoints.put((long) 47, new WayPoint(47, (float) -6.47, (float) -21.69, (float) 0.66, (float) 0.75));
-					this.wayPoints.put((long) 48, new WayPoint(48, (float) -5.91, (float) -1.03, (float) 0.52, (float) 0.85));
-					this.wayPoints.put((long) 49, new WayPoint(49, (float) 6.09, (float) 0.21, (float) -0.04, (float) 1.00));
+					this.wayPoints.put(46L, new WayPoint(46, -6.1f, -28.78f, 0.73f, 0.69f));
+					this.wayPoints.put(47L, new WayPoint(47, -6.47f, -21.69f, 0.66f, 0.75f));
+					this.wayPoints.put(48L, new WayPoint(48, -5.91f, -1.03f, 0.52f, 0.85f));
+					this.wayPoints.put(49L, new WayPoint(49, 6.09f, 0.21f, -0.04f, 1.00f));
 					break;
 				case "U014":
 					this.log.info("Loading waypoints for " + this.currentMap);
